tutorial: Patch up some grammar in section on lens.
The actual content has some problems, but those are being tracked in
sdiehl/wiwinwlh#67.

I tried to salvage the meaning as much as possible, but some of these
sentences are a little complex. :c

I am told that dutch surnames like "Van Laarhoven" should have their first
letter capitalized when used to refer to a person by surname alone. I'm
neither Dutch nor a Dutch speaker, so perhaps this is wrong.

The only controversial thing here is that I swapped "polarizing" for
"staggering". An event or property cannot be "polarizing for some". A
polarizing event or property literally divides all observers into two camps,
the poles, with different (polar) opinions. If something is only partially
polarizing, then it is not polarizing.