Refactor type hole and partial types sections
Type holes are discussed in the `basics` section. The related concept of
partial types is discussed in the `extensions` section. I reorganized
the discussion of both type holes in declarations and wildcards in
signatures to the `basics` section. I leave a short description of how
the `PartialTypeSignatures` extension in the `extensions` section.

I also added the the code used in these sections to `src`.