Move "Type Inference" from extensions to basics
The extensions section begins by categorizing the general types of
extensions. The reader then expects a more detailed introduction to each
extension, which is indeed the general form of the following material.
However, immediately after the extensions section intro, an ectopic
section on type inference begins.

The reason for the type inference section being here isn't at all clear,
so I moved it to the `basics` section.

Following the type inference subsection, is a subsection on the
monomorphism restriction. Here an extension is introduced, but the
material logically groups with type inference. So I merged this section
with the type inference in the `basics` section. In a following commit,
I will add a succinct description of the `NoMonomorphismRestriction`
extension to the extensions section, it will reference the related
material in `basics`.